Start of revamp of image generation
It turns out that revamping image generation also means revamping how Votefinder tracks the current vote state for presentation as HTML, as BBCode *and* as an image. Previously, VotecountFormatter turned all the vote data into a block of BBCode, that was then transmuted into HTML or an image. Now VotecountFormatter creates a game state and then spits out either BBCode or HTML based on that; creating an image from this will be much easier.

The downside is that for some reason I can't quite get the HTML to present nicely on the page. (It still has some BBCode in it, but that shouldn't be affecting things.) This may indicate that the way we're handling custom templates is a problem.